IMA: support to set security.ima instead of signing in runtime
The previous implementation only supports to sign the updated files,
requiring the private key must be deployed on the target.

The new implementation allows the content of signature for each RPM
package to be generated during the build, without requiring the
private key deployed.

Meanwhile, if a private key is really available, the %post script still
prefers using the private key to sign the update files.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>