grub-efi: support to record the last boot option and smart auto-provision
The last boot option is enabled for the regular menu boot option with
savedefault. The one-shot boot option like recovery will be not saved.

In addition, the auto-provision boot option will be automatically
selected when first boot launched. If the auto-provision fails, or clear
after a success, its boot option still appears up but not the default
one.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>