IMA: fix RPM installation failure
If the path of installed file in RPM contains a character requiring
being escaped in shell, e.g, a space, the installation will fail.

The signature list should not use the escaped filename. Otherwise, the
file path in the signature list will be corrupted and points to a
nonexistent place.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>