signing-keys: fix rpm signing failure if both gpg and gpg2 installed
rpm5 always searches gpg2 in higher priority to sign RPM. Hence, if the
signing-keys imports the key with gpg, rpm5 signing will fail for sure
because gpg2 cannot find out the keyring employed by gpg.

This issue only happens if both gpg and gpg2 are installed on a build host.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>