user-key-store-test: adding support for automatically generating RPM signing keys
If the feature user-key-store-test is included, a RPM private key file
will be generated with the keyname = $RPM_GPG_NAME in defined in
user-key-store-test/template.conf.

This patch makes the key generation working with pre-defined RPM keys,
such as Wind River RPM sample key.

All we need to is define a RPM_GPG_NAME, no matter it is a pre-defined key or
a user key.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>