RPM signing enhancement (#93)
* rpm5: grub security patchset from scp8

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>

* rpm_sign: set algorithms for rpm signing

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>

* smartpm: double check package keyid with available pubkey

If a package is not signed, rpm5 will have it self signed with random
key. For these packages, rpm5 always verify it passed, but this is not
what we expect.

We need to install the package signed with installed keys(pubkeys) if
rpm-check-signatures is configured in smart. For other packages, smart
refuses to install.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>

* rpm: make gpg digest algorithm configurable

The default one is SHA1 and it could be overwriten by
RPM_GPG_DIGEST_ALGO.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>

* Set GPG digest algorithm for RPM header signature

We pick up SHA256.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>

* signing-keys: import pubkey for native rpm

Previously native smart doesn't double check imported pubkey with
package's signed keyid if the package is verified pass. But this is not
working now, so we need to import pubkey for native rpm in order to make
sure native smart could install the packages during do_rootfs.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>

* keys: revisit the script for creating user rpm key

We found that rpm5 doesn't support DSA very well, so we change user key
to RSA.

Also we need to export public key, as native rpm will import that key
and narive smart will verify package with it during do_rootfs.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>