user-key-store: fix sel_sign() failure due to using sstate
The commit 9a3ad5e fixes the removal of the ability to use pseudo with
fakeroot. However, the sstate build will break the searching path for
the signaturelet if current build is running in a different location.

Hence, we still need to set LD_LIBRARY_PATH properly and preserve its
value set by pseudo.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>